+++ apps/cli/templates/auth/web/react/react-router/src/components/sign-in-form.tsx
@@ -33,7 +33,7 @@ export default function SignInForm({
             toast.success("Sign in successful");
           },
           onError: (error) => {
-            toast.error(error.error.message);
+            toast.error(error.error.message || error.error.statusText);
           },
         }
       );
@@ -58,7 +58,7 @@ export default function SignInForm({
         onSubmit={(e) => {
           e.preventDefault();
           e.stopPropagation();
-          void form.handleSubmit();
+          form.handleSubmit();
         }}
         className="space-y-4"
       >