+++ apps/web/next.config.mjs
@@ -35,3 +35,7 @@ const config = {
 };
 
 export default withMDX(config);
+
+import { initOpenNextCloudflareForDev } from "@opennextjs/cloudflare";
+
+initOpenNextCloudflareForDev();