+++ apps/web/src/lib/types.ts
@@ -1,27 +1,19 @@
 export type TechCategory =
-	| "core"
-	| "frontend"
+	| "api"
+	| "webFrontend"
+	| "nativeFrontend"
+	| "runtime"
 	| "backend"
 	| "database"
-	| "auth"
 	| "orm"
-	| "router";
-
-export interface TechNode {
-	id: string;
-	type: string;
-	position: { x: number; y: number };
-	data: {
-		label: string;
-		category: TechCategory;
-		description: string;
-		isDefault: boolean;
-		alternatives?: string[];
-		isActive: boolean;
-		group?: TechCategory;
-		isStatic?: boolean;
-	};
-}
+	| "dbSetup"
+	| "webDeploy"
+	| "auth"
+	| "packageManager"
+	| "addons"
+	| "examples"
+	| "git"
+	| "install";
 
 export interface TechEdge {
 	id: string;
@@ -38,6 +30,7 @@ export interface Sponsor {
 		avatarUrl: string;
 		websiteUrl?: string;
 		linkUrl: string;
+		customLogoUrl: string;
 		type: string;
 	};
 	isOneTime: boolean;